# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This a RestAPI for an iGaming platform.",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/distributePrize": {
            "post": {
                "description": "This endpoint triggers a stored procedure to distribute prizes among players based on their rankings in the specified tournament.",
                "tags": [
                    "Prizes"
                ],
                "summary": "Distribute prizes for a tournament",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the tournament",
                        "name": "tournamentID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Prizes distributed successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Tournament ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/rankings": {
            "get": {
                "description": "This endpoint generates a ranking report based on players' account balances in descending order.",
                "tags": [
                    "Rankings"
                ],
                "summary": "Fetch player rankings",
                "responses": {
                    "200": {
                        "description": "List of players with their rankings",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}